backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.2
colored-traceback==0.3.0
colorlog==6.7.0
dill==0.3.7
dj-database-url==1.3.0
Django==3.2.16
django-cors-headers==4.2.0
django-htmlmin==0.11.0
django-phonenumber-field==6.4.0
django-recaptcha==3.0.0
django-tinymce==3.5.0
djangorestframework==3.14.0
EditorConfig==0.12.3
esprima==4.0.1
html5lib==1.1
jsbeautifier==1.14.9
libsass==0.22.0
MarkupSafe==2.1.3
multiprocess==0.70.15
pathos==0.3.1
phonenumbers==8.13.2
Pillow==9.5.0
pox==0.3.3
ppft==1.7.6.7
psycopg2==2.9.5
pyfiglet==0.8.post1
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2022.6
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
typing_extensions==4.5.0
webencodings==0.5.1
whitenoise==6.3.0
